+++ business-app/src/main/java/amaralus/apps/businesappdemo/infrastructure/audit/context/AuditContext.java
@@ -3,29 +3,11 @@ package amaralus.apps.businesappdemo.infrastructure.audit.context;
 import amaralus.apps.businesappdemo.infrastructure.audit.factory.EventFactory;
 import amaralus.apps.businesappdemo.infrastructure.audit.metadata.EntityMetadata;
 
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
+public interface AuditContext {
 
-public class AuditContext {
+    boolean containsMetadata(Class<?> entityClass);
 
-    private final Map<Class<?>, EntityMetadata> entitiesMetadata = new ConcurrentHashMap<>();
-    private final Map<EventFactory.Type, EventFactory> factories = new ConcurrentHashMap<>();
+    EntityMetadata getMetadata(Class<?> entityClass);
 
-    public AuditContext(Map<Class<?>, EntityMetadata> entitiesMetadata, List<EventFactory> eventFactoryList) {
-        this.entitiesMetadata.putAll(entitiesMetadata);
-        eventFactoryList.forEach(factory -> factories.put(factory.getFactoryType(), factory));
-    }
-
-    public EntityMetadata getMetadata(Class<?> entityClass) {
-        return entitiesMetadata.get(entityClass);
-    }
-
-    public boolean containsMetadata(Class<?> entityClass) {
-        return entitiesMetadata.containsKey(entityClass);
-    }
-
-    public EventFactory getEventFactory(EventFactory.Type type) {
-        return factories.get(type);
-    }
+    EventFactory getEventFactory(EventFactory.Type type);
 }
